% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightTriangle.R
\name{rightTriangle}
\alias{rightTriangle}
\title{' Functions to perform Trimmed mean and sides of a  right triangle}
\usage{
rightTriangle(Hypotenuse, opposite, adjacent)
}
\arguments{
\item{Hypotenuse}{Length of the Hypotenuse of a right triangle}

\item{opposite}{Length of the opposite side which is across from a given angle of a right triangle}

\item{adjacent}{Length of the  adjacentside which is next to a given angle of a right triangle.}
}
\value{
side The side that is unknown.
}
\description{
' Functions to perform Trimmed mean and sides of a  right triangle
}
\examples{
rightTriangle(5,4,NA)
rightTriangle(NA,4,3)
}
