% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightTriangle.R
\name{Trimmed_mean}
\alias{Trimmed_mean}
\title{Trimmed mean}
\usage{
Trimmed_mean(x, s, l)
}
\arguments{
\item{x}{a numeric vector thet the Trimmed mean will be calculated}

\item{s}{number of minimum values that is ignored}

\item{l}{number of maximum values that is ignored}
}
\value{
MeanX the mean of the given vector when the requested values are ignored.
}
\description{
The second function will calculate the mean of a numeric vector, ignoring the s smallest value and l largest values.
}
\examples{
Trimmed_mean(c(9,10,11,12),1,2)
Trimmed_mean(c(9,10,11,12),3,1)
}
